refactor: Data관련 함수 수정
